+++ mobile/library/common/jni/jni_interface.cc
@@ -1132,6 +1132,22 @@ Java_io_envoyproxy_envoymobile_engine_JniLibrary_setPreferredNetwork(JNIEnv* env
                                static_cast<envoy_network_t>(network));
 }
 
+extern "C" JNIEXPORT jint JNICALL Java_io_envoyproxy_envoymobile_engine_JniLibrary_setProxySettings(
+    JNIEnv* env,
+    jclass, // class
+    jlong engine, jstring host, jint port) {
+  jni_log("[Envoy]", "setProxySettings");
+
+  const char* native_host = env->GetStringUTFChars(host, nullptr);
+  const uint16_t native_port = static_cast<uint16_t>(port);
+
+  envoy_status_t result =
+      set_proxy_settings(static_cast<envoy_engine_t>(engine), native_host, native_port);
+
+  env->ReleaseStringUTFChars(host, native_host);
+  return result;
+}
+
 bool jvm_cert_is_issued_by_known_root(JNIEnv* env, jobject result) {
   jclass jcls_AndroidCertVerifyResult = env->FindClass("org/chromium/net/AndroidCertVerifyResult");
   jmethodID jmid_isIssuedByKnownRoot =